Rename react-impulse leftovers
