+++ packages/react-impulse/src/batch.ts
@@ -1,4 +1,4 @@
-import { STATIC_SCOPE, type Scope } from "./_Scope"
+import { STATIC_SCOPE, type Scope } from "./scope"
 import { ScopeEmitter } from "./scope-emitter"
 
 /**
