+++ packages/react-impulse/src/index.ts
@@ -1,10 +1,10 @@
-export { type ReadonlyImpulse, Impulse } from "./_Impulse"
+export { type ReadonlyImpulse, Impulse } from "./impulse"
 export type { ReadableImpulse } from "./readable-impulse"
 export type { WritableImpulse } from "./writable-impulse"
 export type { ImpulseOptions } from "./impulse-options"
 export { isImpulse } from "./is-impulse"
 export { isDerivedImpulse } from "./is-derived-impulse"
-export type { Scope } from "./_Scope"
+export type { Scope } from "./scope"
 export { untrack } from "./untrack"
 export { batch } from "./batch"
 export { batch as tap } from "./batch"
