+++ packages/react-impulse/src/is-derived-impulse.ts
@@ -1,7 +1,7 @@
 import { DerivedImpulse } from "./derived-impulse"
-import type { Impulse, ReadonlyImpulse } from "./_Impulse"
+import type { Impulse, ReadonlyImpulse } from "./impulse"
 import { isImpulseFactory } from "./is-impulse-factory"
-import type { Scope } from "./_Scope"
+import type { Scope } from "./scope"
 
 export const isDerivedImpulse = isImpulseFactory(
   (input) => input instanceof DerivedImpulse,
