+++ packages/react-impulse/src/readable-impulse.ts
@@ -1,4 +1,4 @@
-import type { Scope } from "./_Scope"
+import type { Scope } from "./scope"
 
 export interface ReadableImpulse<T> {
   getValue(scope: Scope): T
