+++ packages/react-impulse/src/use-scoped-layout-effect.ts
@@ -1,5 +1,5 @@
 import { type DependencyList, useLayoutEffect } from "./dependencies"
-import type { Scope } from "./_Scope"
+import type { Scope } from "./scope"
 import type { Destructor } from "./destructor"
 import { subscribe } from "./subscribe"
