+++ packages/react-impulse/src/use-scoped-memo.ts
@@ -1,6 +1,6 @@
 import { type DependencyList, useMemo } from "./dependencies"
 import { useCreateScope } from "./use-create-scope"
-import type { Scope } from "./_Scope"
+import type { Scope } from "./scope"
 
 /**
  * The hook is an `Impulse` version of the `React.useMemo` hook.
